+++ messages/it/admin.php
@@ -2,14 +2,8 @@
 
 return [
     '<strong>Dark Mode</strong> module configuration' => 'Configurazione del modulo <strong>Modalità scura</strong>',
-    'Adds a Dark Mode to HumHub' => 'Aggiunge la modalità scura a HumHub',
-    'Current base theme: ' => 'Tema di base corrente ',
     'Dark Mode' => 'Modalità scura',
-    'Dark Theme' => 'Tema scuro',
-    'The stylesheet of the selected theme will be used for the dark mode.' => 'Il foglio di stile del tema selezionato verrà utilizzato per la modalità scura.',
-    'Donate' => '',
-    'GitHub Repository' => '',
-    'Guide for Administrators' => '',
+    'Adds a Dark Mode Switcher to HumHub' => '',
     'Show Button in Top Bar' => '',
     'Users can set their theme preferences also in Account Settings > General.' => '',
 ];